feat(endpoint): centralize `host:port` grammar in aimdb_core for shared use
